% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dyntracer.R
\name{create_dyntracer}
\alias{create_dyntracer}
\title{Create evaldyntracer}
\usage{
create_dyntracer(raw_analysis_dirpath, analysis_flags, verbose)
}
\arguments{
\item{raw_analysis_dirpath}{directory path for storing raw analysis data}

\item{analysis_flags}{flags for enabling/disabling analyses}

\item{verbose}{enable echoing of information}
}
\value{
evaldyntracer
}
\description{
Create evaldyntracer
}
\examples{
create_dyntracer("/tmp", list("eval_expression_analysis" = TRUE), TRUE)

}
