% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destroy_dyntracer.R
\name{destroy_dyntracer}
\alias{destroy_dyntracer}
\title{Destroy evaldyntracer}
\usage{
destroy_dyntracer(evaldyntracer)
}
\arguments{
\item{evaldyntracer}{evaldyntracer object to destroy}
}
\value{
None
}
\description{
Destroy evaldyntracer
}
\examples{
destroy_dyntracer(create_dyntracer("/tmp", list("eval_expression_analysis" = TRUE), TRUE))

}
