% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntrace_evals.R
\name{dyntrace_evals}
\alias{dyntrace_evals}
\title{Trace eval}
\usage{
dyntrace_evals(expr, raw_analysis_dirpath,
  analysis_flags = list(eval_expression_analysis = TRUE), verbose = FALSE)
}
\arguments{
\item{expr}{expression whose evaluation will be traced}

\item{raw_analysis_dirpath}{directory path for storing raw analysis data}

\item{analysis_flags}{flags for enabling/disabling analyses}

\item{verbose}{enable echoing of information}
}
\value{
result of evaluating expr or R_NilValue if an error occurs
}
\description{
Trace eval
}
\examples{
dyntrace_evals({ print(eval("Hello World")) }, "/tmp")

}
