% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaldyntracer.R
\docType{package}
\name{evaldyntracer-package}
\alias{evaldyntracer}
\alias{evaldyntracer-package}
\title{evaldyntracer: a dynamic tracer for eval}
\description{
evaldyntracer is an R package for tracing usage of the eval function. It
intercepts and analyzes the usage and impact of eval in R code. It also
exports the intercepted data for further summarization and analysis.
}
\author{
\strong{Maintainer}: Aviral Goel \email{goel.aviral@gmail.com}

Authors:
\itemize{
  \item Filip Krikava \email{krikava@gmail.com}
  \item Chakshu Goyal \email{chakshugoyal97@gmail.com}
}

}
